refactor: Refactor code to decouple command registry initialization
Extracted command registry initialization to a separate class, CommandRegistry. Modified UserInputHandler to accept a CommandRegistry instance as a dependency for command lookup. Updated Main class to instantiate CommandRegistry and pass it to UserInputHandler. Created the CommandRegistry class to manage the command registry and provide a getCommand method. Adjusted dependencies and class relationships accordingly.

Refs #1
