+++ Learning Progress Tracker (Java)/task/src/tracker/Main.java
@@ -4,7 +4,8 @@ public class Main {
     public static void main(String[] args) {
         OutputProvider outputProvider = new ConsoleOutputProvider();
         InputProvider inputProvider = new SystemInputProvider();
-        UserInputHandler userInputHandler = new UserInputHandler(inputProvider, outputProvider);
+        CommandRegistry commandRegistry = new CommandRegistry(outputProvider);
+        UserInputHandler userInputHandler = new UserInputHandler(inputProvider, commandRegistry);
         LearningProgressTrackerApp app = new LearningProgressTrackerApp(outputProvider, userInputHandler);
 
         app.run();
