+++ Learning Progress Tracker (Java)/task/src/tracker/UserInputHandler.java
@@ -1,30 +1,19 @@
 package tracker;
 
-import java.util.HashMap;
-import java.util.Map;
-
 public class UserInputHandler {
     private final InputProvider inputProvider;
-    private final OutputProvider outputProvider;
-    private final Map<String, Command> commandRegistry;
+    private final CommandRegistry commandRegistry;
 
-    public UserInputHandler(InputProvider inputProvider, OutputProvider outputProvider) {
+    public UserInputHandler(InputProvider inputProvider, CommandRegistry commandRegistry) {
         this.inputProvider = inputProvider;
-        this.outputProvider = outputProvider;
-        this.commandRegistry = new HashMap<>();
-        initializeCommands();
-    }
-
-    private void initializeCommands() {
-        commandRegistry.put("exit", new ExitCommand(outputProvider));
-        // Add other commands to the registry as needed
+        this.commandRegistry = commandRegistry;
     }
 
     public void handleUserInput() {
         while (true) {
             String input = inputProvider.getInput();
 
-            Command command = commandRegistry.get(input);
+            Command command = commandRegistry.getCommand(input);
             if (command != null) {
                 command.execute();
                 break;
